% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nHist.r
\name{nHist}
\alias{nHist}
\title{A Histogram Charting Function}
\usage{
nHist(df, vwr = "pane")
}
\arguments{
\item{df}{Your data frame. It is required. Note: You can chain from dplyr into the chart!}

\item{vwr}{An optional argument. Defaults to 'pane'. Other options include 'dialog' and 'browser'}
}
\value{
An interactive histogram chart
}
\description{
nHIst takes advantage of Shiny Gadgets to make charting basic histograms quick and easy for Birchbox DSAI.
The only required argument is your data frame. The function will chart an interactive histogram that will let you choose which column you want to chart along with the top 'N' you would like to see in the histogram.
}
\details{
There is one optional argument to decide how you would like to view the chart. The default is the rstudio pane. The other options include a pop up dialog or browser.

To close the app, please click "DONE" in the upper right-hand corner!
}
\author{
Matt Negrin
}

